+++ lib/avro/builder/type_factory.rb
@@ -10,6 +10,21 @@ module Avro
 
       private
 
+      # Determine if avro_type_or_name is an existing Type, the name of a builtin
+      # type or a previously defined named type.
+      # If avro_type_or_name is the name of a builtin type, then that type name
+      # is yielded to build the type using a block provided by the caller using,
+      # for example, create_builtin_type or create_and_configure_builtin_type.
+      def type_lookup(avro_type_or_name, namespace = nil)
+        if avro_type_or_name.is_a?(Avro::Builder::Types::Type)
+          avro_type_or_name
+        elsif builtin_type?(avro_type_or_name)
+          yield(avro_type_or_name)
+        else
+          cache.lookup_named_type(avro_type_or_name, namespace)
+        end
+      end
+
       # Return a new Type instance
       def create_builtin_type(avro_type_name, field:, cache:)
         name = avro_type_name.to_s.downcase