+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # avro-builder changelog
 
+## v0.12.0
+- Allow methods for complex and primitive types to be used at the top-level and
+  anywhere that a type name is accepted.
+
 ## v0.11.0
 - Add support for generating schemas that contain logical types. The official
   Ruby `avro` gem does not yet support logical types. The `avro-salsify-fork` gem